+++ packages/asc-ui/src/theme/default/typography.ts
@@ -1,49 +1,45 @@
 import { calculateFluidStyle } from '../../utils'
 import type { Theme } from '../../types'
-import colors from './colors'
 
 const defaultFontFamily = 'Avenir Next, Arial, sans-serif'
 
 const typography: Theme.TypographyInterface = {
   fontFamily: defaultFontFamily,
   h1: {
-    color: colors.tint.level7,
     fontSize: calculateFluidStyle(32, 80),
     fontWeight: 800,
     lineHeight: calculateFluidStyle(40, 88),
+    small: {
+      fontSize: calculateFluidStyle(30, 60),
+      lineHeight: calculateFluidStyle(33, 63),
+    },
   },
   h2: {
-    color: colors.tint.level7,
     fontSize: calculateFluidStyle(28, 56),
     fontWeight: 800,
     lineHeight: calculateFluidStyle(34, 62),
   },
   h3: {
-    color: colors.tint.level7,
     fontSize: calculateFluidStyle(24, 40),
     fontWeight: 800,
     lineHeight: calculateFluidStyle(31, 44),
   },
   h4: {
-    color: colors.tint.level7,
     fontSize: calculateFluidStyle(22, 32),
     fontWeight: 800,
     lineHeight: calculateFluidStyle(31, 38),
   },
   h5: {
-    color: colors.tint.level7,
     fontSize: calculateFluidStyle(18, 24),
     fontWeight: 800,
     lineHeight: calculateFluidStyle(25, 31),
   },
   h6: {
-    color: colors.tint.level7,
     fontSize: calculateFluidStyle(16, 18),
     fontWeight: 800,
     lineHeight: calculateFluidStyle(22, 23),
   },
   a: {
-    color: colors.tint.level7,
     fontSize: '16px',
     fontWeight: 'inherit',
     letterSpacing: 'inherit',
@@ -57,7 +53,6 @@ const typography: Theme.TypographyInterface = {
     },
   },
   p: {
-    color: colors.tint.level7,
     fontSize: '16px',
     fontWeight: 400,
     letterSpacing: 'inherit',
@@ -74,7 +69,6 @@ const typography: Theme.TypographyInterface = {
     },
   },
   li: {
-    color: colors.tint.level7,
     fontSize: '16px',
     fontWeight: 400,
     letterSpacing: 'inherit',
@@ -88,7 +82,6 @@ const typography: Theme.TypographyInterface = {
     },
   },
   em: {
-    color: colors.tint.level7,
     fontSize: '16px',
     fontWeight: 400,
     letterSpacing: 'inherit',
@@ -96,7 +89,6 @@ const typography: Theme.TypographyInterface = {
     marginBottom: '0',
   },
   small: {
-    color: colors.tint.level7,
     fontSize: '14px',
     fontWeight: 400,
     letterSpacing: 'inherit',
@@ -104,7 +96,6 @@ const typography: Theme.TypographyInterface = {
     marginBottom: '0',
   },
   span: {
-    color: colors.tint.level7,
     fontSize: '14px',
     fontWeight: 400,
     letterSpacing: 'inherit',